Prefer class name over bundle id to distinguish plugins
Some people seem to reuse the same bundle identifier for multiple
plugins, so in the end it was not a good idea to use it as a key.

Class names are guaranteed to be unique under the Objective-C runtime.
It seems unlikely that the principle class name changes so often anyway.