Make sure not to pass nil to prevent possible crashes
Some people experience crashes when they close the config window.
The plugin is crashing when it's going to insert a nil key to a
dictionary, which seems to be obtained from the className property.

I still don't recognize cases that the class name taken form an instance
will be nil. Hoping that the defensive nil checks may resolve the issue.